Pickaxes added to shop, anvil added, pickaxes give harvest strength

+++ Reach for the Mines/Assets/_Scripts/Stations/ShopCanvas.cs
@@ -106,7 +106,7 @@ public class ShopCanvas : MonoBehaviour
     {
         if (inventoryItem.empty == false)
         {
-            int goldToAdd = inventoryItem.itemStack.Two * inventoryItem.itemStack.One.value;
+            float goldToAdd = inventoryItem.itemStack.Two * inventoryItem.itemStack.One.value;
             systems.statManager.goldAmount += goldToAdd;
             systems.itemSoldEvent.Invoke();
             systems.inventoryManager.RemoveStack(inventoryItem.itemStack);
